+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Improve: auth.ldap config shown on startup, terminate in case no password is supplied for bind user
 * Add: option [auth] uc_username for uppercase conversion (similar to existing lc_username)
 * Add: option [debug] storage_cache_action for conditional logging
+* Fix: set PRODID on collection upload (instead of vobject is inserting default one)
 
 ## 3.3.1