+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 * Improve: log important module versions on startup
 * Improve: auth.ldap config shown on startup, terminate in case no password is supplied for bind user
 * Add: option [auth] uc_username for uppercase conversion (similar to existing lc_username)
-* Add: option [debug] storage_cache_action for conditional logging
+* Add: option [logging] storage_cache_action_on_debug for conditional logging
 * Fix: set PRODID on collection upload (instead of vobject is inserting default one)
 * Add: option [storage] use_mtime_and_size_for_item_cache for changing cache lookup from SHA256 to mtime_ns + size
 * Fix: buggy cache file content creation on collection upload