+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 3.3.2.dev
+## 3.3.2
 * Fix: debug logging in rights/from_file
 * Add: option [storage] use_cache_subfolder_for_item for storing 'item' cache outside collection-root
 * Fix: ignore empty RRULESET in item